renaming namespace to match the project name
