+++ AspNetCoreVerifiableCredentials/Pages/Index.cshtml.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Verifiable_credentials_DotNet.Pages
+namespace AspNetCoreVerifiableCredentials.Pages
 {
     public class IndexModel : PageModel
     {
