+++ AspNetCoreVerifiableCredentials/Pages/Issuer.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Verifiable_credentials_DotNet.Pages.IssuerModel
+@model AspNetCoreVerifiableCredentials.Pages.IssuerModel
 @{
     ViewData["Title"] = "Issuer";
 }
