+++ AspNetCoreVerifiableCredentials/Pages/Issuer.cshtml.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Verifiable_credentials_DotNet.Pages
+namespace AspNetCoreVerifiableCredentials.Pages
 {
     public class IssuerModel : PageModel
     {
