+++ AspNetCoreVerifiableCredentials/Pages/Verifier.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Verifiable_credentials_DotNet.Pages.VerifierModel
+@model AspNetCoreVerifiableCredentials.Pages.VerifierModel
 @{
     ViewData["Title"] = "Verifier";
 }
