+++ AspNetCoreVerifiableCredentials/Pages/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-@using Verifiable_credentials_DotNet
-@namespace Verifiable_credentials_DotNet.Pages
+@using AspNetCoreVerifiableCredentials
+@namespace AspNetCoreVerifiableCredentials.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
