+++ AspNetCoreVerifiableCredentials/VerifierController.cs
@@ -17,7 +17,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Identity.Web;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Verifiable_credentials_DotNet
+namespace AspNetCoreVerifiableCredentials
 {
     [Route("api/[controller]/[action]")]
     public class VerifierController : Controller
